My First website
This is to showcase my skills and use my ability to use HTML, CSS and JavaScript. (Fresh/New Frontend developer) all transactions are to showcase the functionality and no real transactions will take place on the web page i have not stored the user data to ensure this does not save any information
 
I am sure there will be a bunch more things to work on and any feedback would be great.

I will attempt my python project next but I have to think about what would be best to use for the python project i could attempt to make the python project coincide with the webpage.